chore: Remove GraphQL RulesTable
